+++ state_beacon/lib/src/beacons/derived.dart
@@ -17,10 +17,6 @@ mixin DerivedMixin<T> on ReadableBeacon<T> {
 }
 
 // this is only used internally
-class WritableDerivedBeacon<T> extends ReadableBeacon<T> with DerivedMixin<T> {
+class WritableDerivedBeacon<T> extends WritableBeacon<T> with DerivedMixin<T> {
   WritableDerivedBeacon({super.debugLabel});
-
-  void $forceSet(T newValue) {
-    _setValue(newValue, force: true);
-  }
 }